Don't create new dynamic loaders
Instead of creating a new dynamic lodaer per release, use the original one.

This has the downside that this does NOT work in the editor, as the loader in the editor is monkey patched to get data from the editor, still need to work this out.

But this allows dynamically loaded releases to have state machaines, as the state machine handler was using loadInstance using the original loader..which knows nothing of the dynamically loaded releases